blog/requirements.txt
captcha==0.2.4
certifi==2018.1.18
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.10.8
django-ckeditor==5.4.0
django-crispy-forms==1.7.1
django-filter==1.1.0
django-formtools==2.1
django-guardian==1.4.9
django-js-asset==1.0.0
djangorestframework==3.6.4
DjangoUeditor==1.8.143
future==0.16.0
httplib2==0.9.2
idna==2.6
itypes==1.1.0
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
Pillow==5.0.0
pkg-resources==0.0.0
Pygments==2.2.0
PyMySQL==0.8.0
pytz==2017.3
requests==2.18.4
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
xadmin==0.6.1
